@@ -1,6 +1,6 @@
 package com.emanuelle.easycontacts.ui;
 
-import android.support.v4.app.FragmentActivity;
+import androidx.fragment.app.FragmentActivity;
 import android.os.Bundle;
 
 import com.emanuelle.easycontacts.R;
